Update SVDF unit tests
SVDF int8 is using tflite_micro interpreter for reference data.
To accommodate this the generate script is refactored and split
into smaller files.
