+++ Tests/UnitTest/TestCases/TestData/svdf/biases_data.h
@@ -1,5 +1,6 @@
-// Generated by generate_test_data.py using TFL version 2.6.0 as reference.
+// Generated by generate_test_data.py using tensorflow version 2.10.0 (Keras version 2.10.0).
+// Interpreter from tflite_micro version 0.dev20230817002213-g3bd11ea3 and revision None.
 #pragma once
 #include <stdint.h>
 
-const int32_t svdf_biases[3] = {-53, -125, 86};
+const int32_t svdf_biases[3] = {50, 9, 0};
