+++ Tests/UnitTest/TestCases/TestData/svdf/output_ref_data.h
@@ -1,5 +1,6 @@
-// Generated by generate_test_data.py using TFL version 2.6.0 as reference.
+// Generated by generate_test_data.py using tensorflow version 2.10.0 (Keras version 2.10.0).
+// Interpreter from tflite_micro version 0.dev20230817002213-g3bd11ea3 and revision None.
 #pragma once
 #include <stdint.h>
 
-const int8_t svdf_output_ref[6] = {95, 63, -22, 80, 38, 61};
+const int8_t svdf_output_ref[6] = {73, 52, -128, -128, -16, 61};
