+++ Tests/UnitTest/TestCases/TestData/svdf/weights_feature_data.h
@@ -1,9 +1,10 @@
-// Generated by generate_test_data.py using TFL version 2.6.0 as reference.
+// Generated by generate_test_data.py using tensorflow version 2.10.0 (Keras version 2.10.0).
+// Interpreter from tflite_micro version 0.dev20230817002213-g3bd11ea3 and revision None.
 #pragma once
 #include <stdint.h>
 
-const int8_t svdf_weights_feature[72] = {43,  34,  -69, 80,   -11,  -78, -73, -60, 3,   -98, -111, -40, -102, 15, -99,
-                                         -96, 64,  1,   -120, 117,  94,  -6,  -43, 80,  101, 25,   41,  71,   19, -76,
-                                         55,  -49, -92, 98,   -58,  4,   126, 65,  112, -22, 65,   -86, 38,   64, -86,
-                                         -39, 65,  6,   -97,  -116, 15,  -84, 122, 40,  18,  40,   -78, 84,   63, -35,
-                                         -6,  11,  69,  -45,  51,   80,  -9,  17,  102, 63,  4,    11};
+const int8_t svdf_weights_feature[72] = {
+    29,  -112, 89,  -105, 11, -96, -85, -2,   -24,  114, 75,  -109, 84,  82,  118,  -121, 47,  -11,
+    20,  -37,  111, -87,  22, 71,  14,  99,   93,   -96, -65, 61,   115, -48, 125,  -68,  -62, 59,
+    100, -47,  38,  -23,  30, -54, -64, 0,    -112, -63, -26, 96,   -95, 105, -99,  93,   108, 104,
+    104, -91,  12,  -57,  7,  -13, -13, -124, 70,   -49, 17,  -60,  111, 105, -114, 59,   -81, 55};
