+++ Tests/UnitTest/TestCases/TestData/svdf/weights_time_data.h
@@ -1,16 +1,16 @@
-// Generated by generate_test_data.py using TFL version 2.6.0 as reference.
+// Generated by generate_test_data.py using tensorflow version 2.10.0 (Keras version 2.10.0).
+// Interpreter from tflite_micro version 0.dev20230817002213-g3bd11ea3 and revision None.
 #pragma once
 #include <stdint.h>
 
 const int16_t svdf_weights_time[192] = {
-    -55,  105, 35,  20,   -20,  -123, -38,  92,   88,  13,  119, 14,   -108, -118, -11,  91,   -101, -115,
-    -82,  -6,  -12, 63,   -1,   107,  -24,  2,    67,  85,  -9,  109,  49,   64,   -4,   97,   -110, 18,
-    3,    87,  -52, -56,  -98,  -106, 116,  -29,  100, -93, -46, -8,   -19,  -63,  -58,  -120, 47,   87,
-    -101, 21,  78,  -39,  -77,  -51,  -101, -106, -28, 82,  30,  -35,  16,   105,  -106, 124,  105,  34,
-    -117, 104, 74,  -114, -108, 34,   -111, 32,   8,   -44, -41, -20,  -104, 113,  -63,  116,  -101, 114,
-    61,   125, 103, 20,   -111, 123,  93,   123,  -5,  91,  32,  111,  -116, -1,   35,   69,   -59,  -105,
-    20,   -38, 88,  -3,   65,   -115, 78,   -39,  -49, 121, 23,  -122, 76,   -125, -66,  92,   101,  62,
-    -103, 24,  68,  -104, -82,  -114, -81,  124,  32,  47,  111, 75,   -74,  1,    -90,  61,   106,  123,
-    74,   120, 51,  56,   80,   6,    21,   -100, -35, 119, 11,  -88,  -45,  73,   117,  -79,  -58,  -73,
-    -66,  36,  -54, 11,   -84,  -64,  -87,  -109, 87,  106, -98, 31,   -48,  -47,  -93,  125,  24,   104,
-    -9,   37,  29,  95,   -47,  77,   9,    63,   79,  -92, 82,  22};
+    39,   -11,  13,   -41, -101, -106, -11,  93,   27,   -10,  100,  11,   -128, 13,   125,  -98, 93,  27,  -80, 80,
+    -120, 55,   -37,  27,  24,   41,   -72,  101,  -17,  -21,  116,  -32,  103,  33,   -19,  3,   38,  33,  -7,  -115,
+    -43,  -94,  -77,  -73, -52,  -5,   42,   53,   -22,  105,  -52,  -121, -53,  -49,  -37,  91,  52,  -65, 15,  -58,
+    108,  45,   84,   28,  8,    46,   20,   -101, -6,   -63,  -53,  -105, -74,  -72,  28,   -6,  -30, -88, 84,  67,
+    39,   -92,  -115, -66, 100,  52,   -41,  78,   -41,  115,  42,   42,   -51,  -115, 34,   -89, 11,  35,  71,  64,
+    10,   -30,  -22,  -1,  -23,  -67,  43,   -23,  -110, 43,   42,   105,  109,  74,   -12,  -35, 58,  4,   53,  122,
+    45,   83,   118,  -30, -34,  54,   -111, -60,  28,   -50,  33,   94,   121,  54,   -49,  -32, -56, 27,  95,  29,
+    46,   -86,  -3,   82,  -85,  -119, 90,   -31,  17,   4,    -118, -90,  -116, 65,   -119, -66, -86, 93,  2,   93,
+    6,    -124, 125,  -54, -36,  -85,  39,   -15,  -21,  123,  -44,  -51,  -30,  86,   79,   -16, 70,  -38, -15, 8,
+    33,   -94,  -7,   77,  -17,  -36,  -113, -95,  48,   -121, -92,  -119};
