+++ Tests/UnitTest/TestCases/TestData/svdf_1/input_sequence_data.h
@@ -1,7 +1,8 @@
-// Generated by generate_test_data.py using TFL version 2.6.0 as reference.
+// Generated by generate_test_data.py using tensorflow version 2.10.0 (Keras version 2.10.0).
+// Interpreter from tflite_micro version 0.dev20230817002213-g3bd11ea3 and revision None.
 #pragma once
 #include <stdint.h>
 
-const int8_t svdf_1_input_sequence[42] = {-116, 46,  -102, 125, 27,  -49, -64,  72, 13, 117, -119, 79,  -94,  1,
-                                          -77,  31,  38,   -15, 9,   65,  -102, 30, 10, -30, 16,   -52, 64,   106,
-                                          31,   -59, -78,  -14, -68, 99,  3,    53, 7,  109, -3,   -77, -126, -14};
+const int8_t svdf_1_input_sequence[42] = {-60,  25, 18,  58,  -49,  -118, 95,  33, -72, 25,  -76, 78,  -118, 126,
+                                          -102, 96, -17, -92, -121, -55,  -31, 48, -60, 48,  32,  69,  2,    -68,
+                                          56,   94, -89, 51,  50,   43,   8,   67, 74,  -63, -50, -42, -84,  -43};
