+++ Tests/UnitTest/TestCases/TestData/svdf_1/output_ref_data.h
@@ -1,5 +1,6 @@
-// Generated by generate_test_data.py using TFL version 2.6.0 as reference.
+// Generated by generate_test_data.py using tensorflow version 2.10.0 (Keras version 2.10.0).
+// Interpreter from tflite_micro version 0.dev20230817002213-g3bd11ea3 and revision None.
 #pragma once
 #include <stdint.h>
 
-const int8_t svdf_1_output_ref[15] = {5, 6, 28, 58, -57, 12, -27, 8, -52, 63, -14, 18, -35, 37, -50};
+const int8_t svdf_1_output_ref[15] = {46, -105, -20, -24, -1, 105, 61, 11, 26, -20, -106, 22, -43, -26, 10};
