+++ Tests/UnitTest/TestCases/TestData/svdf_1/weights_feature_data.h
@@ -1,7 +1,8 @@
-// Generated by generate_test_data.py using TFL version 2.6.0 as reference.
+// Generated by generate_test_data.py using tensorflow version 2.10.0 (Keras version 2.10.0).
+// Interpreter from tflite_micro version 0.dev20230817002213-g3bd11ea3 and revision None.
 #pragma once
 #include <stdint.h>
 
-const int8_t svdf_1_weights_feature[35] = {83, 21,  61, 103, 52,   2,   -28, 6,   38,   -10,  25,  -83,
-                                           92, -64, 48, 66,  -4,   57,  40,  108, 39,   -109, -56, -91,
-                                           63, 118, 76, -21, -126, 104, -94, -70, -116, 60,   -82};
+const int8_t svdf_1_weights_feature[35] = {-81,  21, 73,   -122, -91, 43,  -93, 60,  126, -13, 99,   -74,
+                                           -44,  95, -90,  -58,  -50, -29, 90,  65,  25,  -41, -123, 24,
+                                           -100, 23, -113, -88,  48,  -20, 116, -67, 52,  -19, 46};
