+++ Tests/UnitTest/TestCases/TestData/svdf_2/config_data.h
@@ -1,4 +1,5 @@
-// Generated by generate_test_data.py using TFL version 2.6.0 as reference.
+// Generated by generate_test_data.py using tensorflow version 2.10.0 (Keras version 2.10.0).
+// Interpreter from tflite_micro version 0.dev20230817002213-g3bd11ea3 and revision None.
 #pragma once
 #define SVDF_2_MULTIPLIER_IN 1717987072
 #define SVDF_2_MULTIPLIER_OUT 1099511552
