+++ Tests/UnitTest/TestCases/TestData/svdf_2/input_sequence_data.h
@@ -1,7 +1,8 @@
-// Generated by generate_test_data.py using TFL version 2.6.0 as reference.
+// Generated by generate_test_data.py using tensorflow version 2.10.0 (Keras version 2.10.0).
+// Interpreter from tflite_micro version 0.dev20230817002213-g3bd11ea3 and revision None.
 #pragma once
 #include <stdint.h>
 
-const int8_t svdf_2_input_sequence[42] = {
-    29,  81, -38, 17,   -116, 43,   119, -127, 74,   115, 9,   118, 7,   -56,  -53, -14, -98, 60, -128, 10, 28,
-    -18, 12, -28, -126, 87,   -115, -44, -123, -109, -59, -87, -69, 121, -128, -95, -70, 2,   81, -119, 84, -122};
+const int8_t svdf_2_input_sequence[42] = {-83, 20,  39,  85,  7,   70,   -62, 109, 26,  -115, -37, 104, -113, -89,
+                                          -85, -30, 108, 9,   -51, -124, 109, 23,  17,  -58,  58,  89,  -69,  4,
+                                          -16, -2,  -64, 122, 79,  -5,   126, -71, -48, 15,   -28, 111, 81,   75};
