+++ Tests/UnitTest/TestCases/TestData/svdf_2/output_ref_data.h
@@ -1,5 +1,6 @@
-// Generated by generate_test_data.py using TFL version 2.6.0 as reference.
+// Generated by generate_test_data.py using tensorflow version 2.10.0 (Keras version 2.10.0).
+// Interpreter from tflite_micro version 0.dev20230817002213-g3bd11ea3 and revision None.
 #pragma once
 #include <stdint.h>
 
-const int8_t svdf_2_output_ref[15] = {-53, 45, 27, -24, -53, 26, -82, -38, 11, -85, 94, -16, -32, 31, 4};
+const int8_t svdf_2_output_ref[15] = {60, 6, 78, -31, -20, -60, 103, -6, -33, 105, -60, 30, 4, -50, -101};
