+++ Tests/UnitTest/TestCases/TestData/svdf_2/weights_feature_data.h
@@ -1,9 +1,10 @@
-// Generated by generate_test_data.py using TFL version 2.6.0 as reference.
+// Generated by generate_test_data.py using tensorflow version 2.10.0 (Keras version 2.10.0).
+// Interpreter from tflite_micro version 0.dev20230817002213-g3bd11ea3 and revision None.
 #pragma once
 #include <stdint.h>
 
-const int8_t svdf_2_weights_feature[70] = {27,   82,  -108, -127, 85,  3,   -51, 32,  110, -6,  -14, -16,  31,  101,
-                                           -122, 19,  76,   74,   -80, 12,  -22, -17, 10,  -28, 55,  109,  2,   -107,
-                                           -4,   72,  -65,  -59,  36,  -69, 105, -97, 25,  38,  110, -121, -88, -126,
-                                           -14,  16,  -88,  -66,  3,   -93, 69,  -64, 44,  103, 95,  -95,  68,  -46,
-                                           106,  -31, -63,  23,   -38, 36,  -95, -43, 93,  77,  91,  -26,  33,  59};
+const int8_t svdf_2_weights_feature[70] = {89,  -63,  93,  -59, -108, -25, 74,  -63,  104,  24,   126, -56, 100, 64,
+                                           88,  61,   -6,  76,  111,  39,  73,  -79,  37,   -2,   10,  -99, 31,  -125,
+                                           53,  -94,  114, -86, 119,  -56, 109, -124, 116,  -113, 38,  -63, 28,  -32,
+                                           -26, -61,  -85, 66,  109,  -51, 39,  -112, -123, -22,  62,  62,  102, -62,
+                                           65,  -107, -75, 107, -86,  53,  -50, 77,   62,   -29,  -51, 109, 43,  -95};
