+++ Tests/UnitTest/TestCases/TestData/svdf_2/weights_time_data.h
@@ -1,6 +1,7 @@
-// Generated by generate_test_data.py using TFL version 2.6.0 as reference.
+// Generated by generate_test_data.py using tensorflow version 2.10.0 (Keras version 2.10.0).
+// Interpreter from tflite_micro version 0.dev20230817002213-g3bd11ea3 and revision None.
 #pragma once
 #include <stdint.h>
 
-const int16_t svdf_2_weights_time[20] = {-31, -88, -10, -72, -119, -6, -70, 63,  -10, 93,
-                                         5,   42,  -6,  22,  6,    51, 37,  -38, 5,   117};
+const int16_t svdf_2_weights_time[20] = {-99, 75, 86, 3,   78,  18,  -70, -103, -49, 110,
+                                         124, 7,  29, -52, -10, -67, 1,   18,   111, -86};
