+++ Tests/UnitTest/TestCases/TestData/svdf_3/input_sequence_data.h
@@ -1,7 +1,8 @@
-// Generated by generate_test_data.py using TFL version 2.6.0 as reference.
+// Generated by generate_test_data.py using tensorflow version 2.10.0 (Keras version 2.10.0).
+// Interpreter from tflite_micro version 0.dev20230817002213-g3bd11ea3 and revision None.
 #pragma once
 #include <stdint.h>
 
-const int8_t svdf_3_input_sequence[40] = {-64, -83, 38, 1,   119, -124, -53,  -63, 52,   -23, -22, -122, -79, -22,
-                                          80,  87,  14, -3,  107, -67,  57,   -2,  -104, 97,  62,  -26,  101, -126,
-                                          -53, 10,  23, -32, 28,  -68,  -108, 100, -8,   117, -72, -26};
+const int8_t svdf_3_input_sequence[40] = {-78, 32, 72,  -36, -40,  21,  122, -22,  -76, 114,  -122, -115, -102, 55,
+                                          70,  87, 120, -51, 124,  -39, -8,  -116, -57, -100, -43,  -18,  13,   30,
+                                          -43, 76, 80,  29,  -111, -60, 71,  103,  75,  -84,  -79,  -35};
