+++ Tests/UnitTest/TestCases/TestData/svdf_3/output_ref_data.h
@@ -1,5 +1,6 @@
-// Generated by generate_test_data.py using TFL version 2.6.0 as reference.
+// Generated by generate_test_data.py using tensorflow version 2.10.0 (Keras version 2.10.0).
+// Interpreter from tflite_micro version 0.dev20230817002213-g3bd11ea3 and revision None.
 #pragma once
 #include <stdint.h>
 
-const int8_t svdf_3_output_ref[12] = {20, 74, 56, 65, -23, 57, -49, -128, 2, 109, 72, -13};
+const int8_t svdf_3_output_ref[12] = {43, -46, 71, 45, -89, 71, -105, 59, 111, 127, 43, 127};
