+++ Tests/UnitTest/TestCases/TestData/svdf_3/weights_feature_data.h
@@ -1,17 +1,18 @@
-// Generated by generate_test_data.py using TFL version 2.6.0 as reference.
+// Generated by generate_test_data.py using tensorflow version 2.10.0 (Keras version 2.10.0).
+// Interpreter from tflite_micro version 0.dev20230817002213-g3bd11ea3 and revision None.
 #pragma once
 #include <stdint.h>
 
 const int8_t svdf_3_weights_feature[240] = {
-    -71,  -110, 122,  -115, -52,  -35,  -58, -81,  12,   -127, 88,   -67,  -114, 7,    -69,  73,  113,  -80,  -55,  -72,
-    -113, 45,   -13,  116,  -121, 38,   31,  -10,  56,   47,   -43,  11,   96,   106,  44,   -80, 48,   -37,  46,   -87,
-    -88,  -66,  63,   86,   56,   76,   37,  47,   -16,  77,   -68,  16,   -48,  42,   -107, 104, -103, 8,    -72,  -24,
-    -68,  -58,  -127, -75,  -121, 116,  69,  -42,  -32,  -38,  71,   57,   -116, -3,   -42,  -63, 10,   -3,   71,   56,
-    -24,  46,   -125, 102,  35,   -111, -79, 110,  -37,  -10,  -127, -103, -34,  44,   4,    -15, -12,  -35,  -121, -66,
-    -38,  -28,  -17,  -33,  -102, -127, -4,  50,   82,   93,   26,   88,   9,    59,   -119, 33,  26,   -117, -24,  102,
-    11,   72,   39,   -123, 20,   -29,  -53, 23,   12,   32,   64,   -3,   -81,  -54,  66,   67,  40,   -106, 126,  -3,
-    69,   67,   62,   102,  38,   104,  -78, -107, -106, 100,  64,   57,   -65,  26,   -94,  63,  -62,  28,   -80,  101,
-    -48,  -66,  -1,   -63,  1,    -85,  98,  66,   46,   -70,  50,   99,   -59,  -22,  -74,  -43, -75,  -16,  30,   90,
-    93,   68,   -53,  11,   46,   -92,  -12, -118, -45,  -62,  48,   -76,  35,   82,   73,   82,  -100, -109, 48,   62,
-    103,  110,  -5,   -18,  110,  0,    121, -70,  54,   -75,  116,  126,  16,   -32,  -126, -34, 45,   -100, 112,  -63,
-    8,    -100, 55,   -2,   -33,  -127, 97,  82,   -89,  8,    110,  -4,   -105, -110, -42,  -24, -35,  -34,  -34,  8};
+    -63,  -104, 46,   7,    97,   120, -35, 109,  78,   43,   -26,  42,   -57, -89,  -39, 111,  -31,  -16,  85,   102,
+    -16,  44,   -3,   -110, 5,    59,  -78, -89,  -29,  -94,  -19,  -23,  -95, -54,  -58, 120,  -89,  -101, 89,   -51,
+    9,    93,   -72,  98,   21,   107, 92,  116,  28,   16,   -75,  -59,  -99, 63,   -39, 107,  -76,  -110, -75,  -97,
+    -19,  44,   -70,  77,   116,  -66, 110, -35,  -86,  8,    57,   2,    9,   47,   38,  98,   22,   37,   32,   2,
+    80,   -49,  43,   58,   43,   41,  38,  -98,  -120, -120, 85,   -39,  -12, 68,   26,  -36,  -97,  93,   -34,  21,
+    -107, -3,   -114, -56,  -28,  37,  -98, 92,   105,  -120, -105, -12,  -29, -100, -56, -73,  85,   -15,  -47,  54,
+    38,   -90,  -38,  -91,  -96,  11,  126, -121, -36,  -20,  86,   -37,  114, -7,   2,   27,   27,   -128, 94,   -47,
+    75,   -34,  93,   -125, 82,   -54, 51,  92,   -91,  53,   112,  -19,  9,   -56,  0,   68,   -45,  -22,  56,   -54,
+    95,   62,   37,   89,   18,   -44, -80, 86,   11,   -82,  122,  -113, 61,  -45,  -72, -125, -111, 118,  124,  -35,
+    70,   -15,  -100, -116, -79,  -8,  -5,  -8,   -27,  37,   -8,   84,   -36, 61,   114, 123,  30,   -30,  -104, 23,
+    104,  56,   -29,  -56,  -50,  99,  22,  -127, -56,  126,  41,   103,  49,  -118, 95,  87,   -11,  0,    92,   17,
+    -1,   -37,  17,   2,    -108, -28, 14,  97,   -17,  87,   -29,  3,    0,   15,   11,  100,  -52,  -85,  125,  -122};
