+++ Tests/UnitTest/TestCases/TestData/svdf_3/weights_time_data.h
@@ -1,6 +1,7 @@
-// Generated by generate_test_data.py using TFL version 2.6.0 as reference.
+// Generated by generate_test_data.py using tensorflow version 2.10.0 (Keras version 2.10.0).
+// Interpreter from tflite_micro version 0.dev20230817002213-g3bd11ea3 and revision None.
 #pragma once
 #include <stdint.h>
 
-const int16_t svdf_3_weights_time[24] = {-17,  -93, -78, -65,  46,   109, -70, -68, -41, -47, 2,   -108,
-                                         -108, -42, 69,  -103, -127, 73,  100, 113, -52, 98,  -74, -14};
+const int16_t svdf_3_weights_time[24] = {39,  109, -70, -57, 109, -78, 91, -68, 24, 118, -104, 105,
+                                         -75, -87, 17,  65,  -84, 7,   -9, 88,  -4, 75,  73,   91};
