+++ Tests/UnitTest/TestCases/TestData/svdf_int8/config_data.h
@@ -1,12 +1,12 @@
 // Generated by generate_test_data.py using tensorflow version 2.10.0 (Keras version 2.10.0).
-// Interpreter from tensorflow version 2.10.0 and revision v2.10.0-rc3-6-g359c3cdfc5f.
+// Interpreter from tflite_micro version 0.dev20230817002213-g3bd11ea3 and revision None.
 #pragma once
 #define SVDF_INT8_MULTIPLIER_IN 1717987072
 #define SVDF_INT8_MULTIPLIER_OUT 1099511552
 #define SVDF_INT8_SHIFT_1 -3
 #define SVDF_INT8_SHIFT_2 -11
-#define SVDF_INT8_IN_ACTIVATION_MIN -32768
-#define SVDF_INT8_IN_ACTIVATION_MAX 32767
+#define SVDF_INT8_IN_ACTIVATION_MIN -128
+#define SVDF_INT8_IN_ACTIVATION_MAX 127
 #define SVDF_INT8_RANK 1
 #define SVDF_INT8_FEATURE_BATCHES 12
 #define SVDF_INT8_TIME_BATCHES 2
