+++ Tests/UnitTest/TestCases/TestData/svdf_int8/test_data.h
@@ -1,8 +1,9 @@
 // Generated by generate_test_data.py using tensorflow version 2.10.0 (Keras version 2.10.0).
-// Interpreter from tensorflow version 2.10.0 and revision v2.10.0-rc3-6-g359c3cdfc5f.
+// Interpreter from tflite_micro version 0.dev20230817002213-g3bd11ea3 and revision None.
 #include "biases_data.h"
 #include "config_data.h"
 #include "input_sequence_data.h"
+#include "output_ref_data.h"
 #include "state_data.h"
 #include "weights_feature_data.h"
 #include "weights_time_data.h"
